ml_pipeline/requirements.txt
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
pyyaml==6.0.1
python-dotenv==1.0.0
supabase==2.4.0
click==8.1.7
pytest==7.4.3
pytest-cov==4.1.0
